add properties "CaloricContentIn100gr" in classes add automatic calculation for property "CaloricContent"

+++ Task 1/SchefPovar/Interfaces/IComponent.cs
@@ -9,7 +9,7 @@ namespace SchefPovar
     public interface IComponent
     {
         string Name { get; }
-        int CaloricContent { get; }
+        double CaloricContent { get; }
 
         string AllProperties{get;}
